+++ pdeque_test.go
@@ -153,6 +153,41 @@ var _ = Describe("type PDeque", func() {
 		})
 	})
 
+	Describe("func Contains()", func() {
+		It("returns true if the element is in the queue", func() {
+			e := queue.Push(0)
+			Expect(queue.Contains(e)).To(BeTrue())
+		})
+
+		It("returns true if the element is in a different queue", func() {
+			var q PDeque
+			e := q.Push(0)
+			Expect(queue.Contains(e)).To(BeFalse())
+		})
+
+		It("returns false if the element has been removed from the queue", func() {
+			e := queue.Push(0)
+			queue.Pop()
+			Expect(queue.Contains(e)).To(BeFalse())
+		})
+	})
+
+	Describe("func IsFront()", func() {
+		It("returns false if the element is not in the queue", func() {
+			e := queue.Push(0)
+			queue.Pop()
+			Expect(queue.IsFront(e)).To(BeFalse())
+		})
+	})
+
+	Describe("func IsBack()", func() {
+		It("returns false if the element is not in the queue", func() {
+			e := queue.Push(0)
+			queue.Pop()
+			Expect(queue.IsBack(e)).To(BeFalse())
+		})
+	})
+
 	Describe("func Update()", func() {
 		It("repairs the queue order", func() {
 			queue.Push(10)
@@ -162,6 +197,15 @@ var _ = Describe("type PDeque", func() {
 			e.Value = 5
 			queue.Update(e)
 			Expect(queue.IsFront(e)).To(BeTrue())
+		})
+
+		It("panics if the element is not on the queue", func() {
+			e := queue.Push(0)
+			queue.Pop()
+
+			Expect(func() {
+				queue.Update(e)
+			}).To(PanicWith("element is not on the queue"))
 		})
 	})
 
@@ -180,6 +224,12 @@ var _ = Describe("type PDeque", func() {
 			v, ok = queue.Pop()
 			Expect(ok).To(BeTrue())
 			Expect(v).To(Equal(30))
+		})
+
+		It("does not panic if the element is not in the queue", func() {
+			e := queue.Push(0)
+			queue.Pop()
+			queue.Remove(e)
 		})
 	})